* Integer literal now back to int from long * Errors cleaned up and more detailed * HashMap now correctly initialized * Added increment and decrement methods to execution * Error handling in execution * Exit codes enum to handle various exits

+++ src/App/Interpreter/Node/Nodes/IntegerLiteral.java
@@ -3,13 +3,13 @@ package App.Interpreter.Node.Nodes;
 import App.Interpreter.Node.*;
 
 public class IntegerLiteral implements Node {
-  private final long value;
+  private final int value;
 
-  public IntegerLiteral(long value) {
+  public IntegerLiteral(int value) {
     this.value = value;
   }
 
-  public long getValue() {
+  public int getValue() {
     return value;
   }
